refactor(loop): 集中注册生命周期消费者
