+++ agent/looping/lifecycle_consumers.py
@@ -1,0 +1,133 @@
+from __future__ import annotations
+
+import json
+from collections.abc import Mapping
+from typing import TYPE_CHECKING, Protocol, cast, runtime_checkable
+
+from bus.events_lifecycle import PostTurnScheduled, TurnPersisted
+
+if TYPE_CHECKING:
+    from agent.looping.consolidation import ConsolidationService
+    from agent.looping.ports import ObservabilityServices
+    from bus.event_bus import EventBus
+
+
+@runtime_checkable
+class _ObserveWriter(Protocol):
+    def emit(self, event: object) -> None: ...
+
+
+def register_post_turn_consumers(
+    *,
+    event_bus: "EventBus",
+    consolidation: "ConsolidationService",
+) -> None:
+    async def _refresh_recent_context(event: PostTurnScheduled) -> None:
+        await consolidation.refresh_recent_turns(session=event.session)
+
+    event_bus.on(PostTurnScheduled, _refresh_recent_context)
+
+
+def register_observe_trace_consumers(
+    *,
+    event_bus: "EventBus",
+    trace: "ObservabilityServices",
+) -> None:
+    writer = trace.observe_writer
+    if not isinstance(writer, _ObserveWriter):
+        return
+
+    def _observe_turn_persisted(event: TurnPersisted) -> None:
+        _emit_turn_trace(writer, event)
+
+    event_bus.on(TurnPersisted, _observe_turn_persisted)
+
+
+def _emit_turn_trace(writer: _ObserveWriter, event: TurnPersisted) -> None:
+    from core.observe.events import TurnTrace as TurnTraceEvent
+
+    post_reply_budget = event.post_reply_budget
+    react_stats = event.react_stats
+    tool_chain = event.tool_chain
+    tool_chain_json = (
+        json.dumps(_slim_tool_chain(tool_chain), ensure_ascii=False)
+        if tool_chain
+        else None
+    )
+    writer.emit(
+        TurnTraceEvent(
+            source="agent",
+            session_key=event.session_key,
+            user_msg=event.user_message,
+            llm_output=event.assistant_response,
+            raw_llm_output=event.raw_reply,
+            meme_tag=event.meme_tag,
+            meme_media_count=event.meme_media_count,
+            tool_calls=_slim_tool_calls(tool_chain),
+            tool_chain_json=tool_chain_json,
+            history_window=post_reply_budget.get("history_window"),
+            history_messages=post_reply_budget.get("history_messages"),
+            history_chars=post_reply_budget.get("history_chars"),
+            history_tokens=post_reply_budget.get("history_tokens"),
+            prompt_tokens=post_reply_budget.get("prompt_tokens"),
+            next_turn_baseline_tokens=post_reply_budget.get(
+                "next_turn_baseline_tokens"
+            ),
+            react_iteration_count=react_stats.get("iteration_count"),
+            react_input_sum_tokens=react_stats.get("turn_input_sum_tokens"),
+            react_input_peak_tokens=react_stats.get("turn_input_peak_tokens"),
+            react_final_input_tokens=react_stats.get("final_call_input_tokens"),
+            react_cache_prompt_tokens=react_stats.get("cache_prompt_tokens"),
+            react_cache_hit_tokens=react_stats.get("cache_hit_tokens"),
+        )
+    )
+    if event.retrieval_raw is not None:
+        writer.emit(event.retrieval_raw)
+
+
+def _slim_tool_calls(tool_chain: list[dict[str, object]]) -> list[dict[str, str]]:
+    return [
+        {
+            "name": str(call.get("name", "")),
+            "args": str(call.get("arguments", ""))[:300],
+            "result": str(call.get("result", ""))[:500],
+        }
+        for group in tool_chain
+        for call in _group_calls(group)
+    ]
+
+
+def _slim_tool_chain(tool_chain: list[dict[str, object]]) -> list[dict[str, object]]:
+    return [
+        {
+            "text": str(group.get("text") or ""),
+            "calls": [
+                {
+                    "name": str(call.get("name", "")),
+                    "args": str(call.get("arguments", ""))[:800],
+                    "result": str(call.get("result", ""))[:1200],
+                }
+                for call in _group_calls(group)
+            ],
+        }
+        for group in tool_chain
+    ]
+
+
+def _group_calls(group: dict[str, object]) -> list[dict[str, object]]:
+    calls = group.get("calls")
+    if not isinstance(calls, list):
+        return []
+    raw_calls = cast(list[object], calls)
+    out: list[dict[str, object]] = []
+    for call in raw_calls:
+        if isinstance(call, Mapping):
+            mapping = cast(Mapping[object, object], call)
+            out.append(
+                {
+                    str(key): value
+                    for key, value in mapping.items()
+                    if isinstance(key, str)
+                }
+            )
+    return out
